src/pyproject.toml
[tool.poetry]
name = "Fitness Studio"
version = "0.1.0"
description = " API Service"
authors = ["temp"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.18.3"
fastapi = "^0.98.0"
pydantic= "^1.10.13"
fastapi-event = "^0.1.3"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
websockets = "^12.0"
load-dotenv = "^0.1.0"
pyjwt = "^2.8.0"
sqlalchemy = {version = "1.4.40", extras = ["asyncio"]}
sqlalchemy-events = "^0.1.0"

[tool.poetry.dev-dependencies]
behave = "^1.2.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"